final decission bowtie settings: --very-fast
Comparing bowtie settings, the increase in reads mapped with settings --very-sensitive -N 1 is low. Also, the assembly and subsequent annotation and re-filtering using cat and again bowtie is thought to be highly efficient and more robust // less biased than 'simple homology' to a reference genome. 
To keep the pipeline from unnecessary calculating, I chose to keep the initial filtering step in but with very strict settings: --very-fast.